# app/api/aiden-mesh/discover/route.ts
import { NextRequest, NextResponse } from 'next/server'

// Peer discovery endpoint for AIDEN mesh
// Allows subsystems to find and connect with each other

interface SubsystemInfo {
  id: string
  name: string
  type: 'swarm-agent' | 'compliance' | 'blockchain' | 'multimodal' | 'quantum' | 'orchestrator' | 'workload'
  status: 'online' | 'offline' | 'degraded'
  capabilities: string[]
  endpoints: string[]
  lastHeartbeat: string
  metrics: any
  registeredAt: string
}

// Mock subsystems for discovery (in production, would query mesh state)
const registeredSubsystems: SubsystemInfo[] = [
  {
    id: 'swarm-core-001',
    name: 'LLM Swarm Orchestrator',
    type: 'swarm-agent',
    status: 'online',
    capabilities: ['task-execution', 'agent-coordination', 'dna-lang-runtime'],
    endpoints: ['/api/swarm/agents', '/api/swarm/agents/{id}/execute'],
    lastHeartbeat: new Date().toISOString(),
    metrics: { lambda: 0.85, gamma: 0.12, phi: 0.72, w2: 0.15 },
    registeredAt: new Date(Date.now() - 3600000).toISOString()
  },
  {
    id: 'compliance-core-001',
    name: 'Compliance & Audit Core',
    type: 'compliance',
    status: 'online',
    capabilities: ['compliance-check', 'audit-logging', 'fraud-detection'],
    endpoints: ['/api/compliance/check', '/api/compliance/audit-log', '/api/fraud/detect'],
    lastHeartbeat: new Date().toISOString(),
    metrics: { lambda: 0.92, gamma: 0.08, phi: 0.88, w2: 0.10 },
    registeredAt: new Date(Date.now() - 7200000).toISOString()
  },
  {
    id: 'blockchain-exec-001',
    name: 'QuantumCoin Blockchain Executor',
    type: 'blockchain',
    status: 'online',
    capabilities: ['transaction-execution', 'smart-contracts', 'quantum-verification'],
    endpoints: ['/api/blockchain/execute', '/api/blockchain/verify'],
    lastHeartbeat: new Date().toISOString(),
    metrics: { lambda: 0.78, gamma: 0.15, phi: 0.65, w2: 0.18 },
    registeredAt: new Date(Date.now() - 5400000).toISOString()
  },
  {
    id: 'multimodal-chat-001',
    name: 'Enhanced Multimodal Chat',
    type: 'multimodal',
    status: 'online',
    capabilities: ['file-upload', 'camera-capture', 'screen-sharing', 'web-browsing', 'cloud-storage'],
    endpoints: ['/api/upload', '/api/cloud/picker', '/api/web/screenshot', '/api/chat'],
    lastHeartbeat: new Date().toISOString(),
    metrics: { lambda: 0.88, gamma: 0.11, phi: 0.76, w2: 0.13 },
    registeredAt: new Date(Date.now() - 1800000).toISOString()
  },
  {
    id: 'quantum-mesh-001',
    name: 'IBM Quantum Integration',
    type: 'quantum',
    status: 'online',
    capabilities: ['quantum-execution', 'circuit-optimization', 'backend-selection', 'lambda-phi-calculation'],
    endpoints: ['/api/quantum/backends', '/api/quantum/status', '/api/quantum/submit-job'],
    lastHeartbeat: new Date().toISOString(),
    metrics: { lambda: 0.95, gamma: 0.06, phi: 0.91, w2: 0.08 },
    registeredAt: new Date(Date.now() - 10800000).toISOString()
  },
  {
    id: 'orchestrator-001',
    name: 'Quantum Swarm Orchestrator',
    type: 'orchestrator',
    status: 'online',
    capabilities: ['watson-ai', 'agent-management', 'permission-system', 'activity-logging'],
    endpoints: ['/api/orchestrator/profile', '/api/orchestrator/agents', '/api/orchestrator/permissions', '/api/orchestrator/activities'],
    lastHeartbeat: new Date().toISOString(),
    metrics: { lambda: 0.89, gamma: 0.10, phi: 0.80, w2: 0.12 },
    registeredAt: new Date(Date.now() - 9000000).toISOString()
  },
  {
    id: 'workload-analytics-001',
    name: 'Quantum Workload Analytics',
    type: 'workload',
    status: 'online',
    capabilities: ['benchmark-tracking', 'performance-monitoring', 'cost-analysis'],
    endpoints: ['/api/benchmarks'],
    lastHeartbeat: new Date().toISOString(),
    metrics: { lambda: 0.82, gamma: 0.14, phi: 0.70, w2: 0.16 },
    registeredAt: new Date(Date.now() - 4500000).toISOString()
  }
]

// GET /api/aiden-mesh/discover - Discover peer subsystems
export async function GET(request: NextRequest) {
  try {
    const { searchParams } = new URL(request.url)
    const typeFilter = searchParams.get('type')
    const capabilityFilter = searchParams.get('capability')

    let results = [...registeredSubsystems]

    // Filter by type
    if (typeFilter) {
      results = results.filter(s => s.type === typeFilter)
    }

    // Filter by capability
    if (capabilityFilter) {
      results = results.filter(s => s.capabilities.includes(capabilityFilter))
    }

    // Only return online subsystems
    results = results.filter(s => s.status === 'online')

    // Sort by lambda (highest first) - most coherent subsystems first
    results.sort((a, b) => (b.metrics.lambda || 0) - (a.metrics.lambda || 0))

    console.log(`[AIDEN Mesh] Peer discovery: ${results.length} subsystems found (type=${typeFilter}, capability=${capabilityFilter})`)

    return NextResponse.json(results)

  } catch (error: any) {
    console.error('Error during peer discovery:', error)
    return NextResponse.json(
      { error: error.message || 'Failed to discover peers' },
      { status: 500 }
    )
  }
}

// POST /api/aiden-mesh/discover - Announce capability
export async function POST(request: NextRequest) {
  try {
    const body = await request.json()

    // Subsystem announces a new capability or status change
    if (!body.subsystemId) {
      return NextResponse.json(
        { error: 'subsystemId required' },
        { status: 400 }
      )
    }

    console.log(`[AIDEN Mesh] Capability announcement from ${body.subsystemId}:`, body)

    // In production, would update mesh state and notify interested peers
    return NextResponse.json({
      success: true,
      message: 'Capability announcement received',
      timestamp: new Date().toISOString()
    })

  } catch (error: any) {
    console.error('Error processing capability announcement:', error)
    return NextResponse.json(
      { error: error.message || 'Failed to process announcement' },
      { status: 500 }
    )
  }
}